fix: ensure stable setState/dispatch identity across renders
Critical fixes addressing CodeRabbit feedback:

**Fix setter identity thrashing (React contract violation)**
- useState and useReducer now return stable function references
- Capture fiber at hook initialization instead of creating new closures
- Prevents infinite loops from dependency array invalidation
- Fixes effects/memos that depend on setState/dispatch

**Refactor effect queue into legacy scheduler**
- Move effectQueue to dedicated src/hooks/legacyScheduler.ts module
- Cleaner separation of concerns
- Add resetEffectQueueForTests() to prevent test pollution
- Maintain backward compatibility with existing code

This fixes major performance issues where every render created new
setState/dispatch functions, causing unnecessary re-renders and
breaking React's stability guarantees.

All 520 tests passing, zero type errors.
